refactor(cli): organize commands into dedicated packages
