+++ cmd/semver/bumpcmd/next.go
@@ -1,0 +1,177 @@
+package bumpcmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+
+	"github.com/indaco/semver-cli/internal/clix"
+	"github.com/indaco/semver-cli/internal/config"
+	"github.com/indaco/semver-cli/internal/hooks"
+	commitparser "github.com/indaco/semver-cli/internal/plugins/commit-parser"
+	"github.com/indaco/semver-cli/internal/plugins/commit-parser/gitlog"
+	"github.com/indaco/semver-cli/internal/semver"
+	"github.com/urfave/cli/v3"
+)
+
+var tryInferBumpTypeFromCommitParserPluginFn = tryInferBumpTypeFromCommitParserPlugin
+
+// nextCmd returns the "next" subcommand.
+func nextCmd(cfg *config.Config) *cli.Command {
+	return &cli.Command{
+		Name:  "next",
+		Usage: "Smart bump logic (e.g. promote pre-release or bump patch)",
+		UsageText: `semver bump next [--label patch|minor|major] [--meta data] [--preserve-meta] [--since ref] [--until ref] [--no-infer]
+
+By default, semver tries to infer the bump type from recent commit messages using the built-in commit-parser plugin.
+You can override this behavior with the --label flag, disable it explicitly with --no-infer, or disable the plugin via the config file (.semver.yaml).`,
+		Flags: []cli.Flag{
+			&cli.StringFlag{
+				Name:  "label",
+				Usage: "Optional bump label override (patch, minor, major)",
+			},
+			&cli.StringFlag{
+				Name:  "meta",
+				Usage: "Set build metadata (e.g. 'ci.123')",
+			},
+			&cli.BoolFlag{
+				Name:  "preserve-meta",
+				Usage: "Preserve existing build metadata instead of clearing it",
+			},
+			&cli.StringFlag{
+				Name:  "since",
+				Usage: "Start commit/tag for bump inference (default: last tag or HEAD~10)",
+			},
+			&cli.StringFlag{
+				Name:  "until",
+				Usage: "End commit/tag for bump inference (default: HEAD)",
+			},
+			&cli.BoolFlag{
+				Name:  "no-infer",
+				Usage: "Disable bump inference from commit messages (overrides config)",
+			},
+			&cli.BoolFlag{
+				Name:  "hook-only",
+				Usage: "Only run pre-release hooks, do not modify the version",
+			},
+			&cli.BoolFlag{
+				Name:  "skip-hooks",
+				Usage: "Skip pre-release hooks",
+			},
+		},
+		Action: func(ctx context.Context, cmd *cli.Command) error {
+			return runBumpNext(cfg, cmd)
+		},
+	}
+}
+
+// runBumpNext performs smart bumping (e.g. promote, patch, infer).
+func runBumpNext(cfg *config.Config, cmd *cli.Command) error {
+	path := cmd.String("path")
+	label := cmd.String("label")
+	meta := cmd.String("meta")
+	since := cmd.String("since")
+	until := cmd.String("until")
+	isPreserveMeta := cmd.Bool("preserve-meta")
+	isNoInferFlag := cmd.Bool("no-infer")
+	isSkipHooks := cmd.Bool("skip-hooks")
+
+	disableInfer := isNoInferFlag || (cfg != nil && cfg.Plugins != nil && !cfg.Plugins.CommitParser)
+
+	if _, err := clix.FromCommand(cmd); err != nil {
+		return err
+	}
+
+	current, err := semver.ReadVersion(path)
+	if err != nil {
+		return fmt.Errorf("failed to read version: %w", err)
+	}
+
+	if err := hooks.RunPreReleaseHooks(isSkipHooks); err != nil {
+		return err
+	}
+
+	var next semver.SemVersion
+
+	switch label {
+	case "patch", "minor", "major":
+		next, err = semver.BumpByLabelFunc(current, label)
+		if err != nil {
+			return fmt.Errorf("failed to bump version with label: %w", err)
+		}
+	case "":
+		if !disableInfer {
+			inferred := tryInferBumpTypeFromCommitParserPluginFn(since, until)
+			if inferred != "" {
+				fmt.Fprintf(os.Stderr, "🔍 Inferred bump type: %s\n", inferred)
+
+				if current.PreRelease != "" {
+					next = promotePreRelease(current, isPreserveMeta)
+				} else {
+					next, err = semver.BumpByLabelFunc(current, inferred)
+					if err != nil {
+						return fmt.Errorf("failed to bump inferred version: %w", err)
+					}
+				}
+				break
+			}
+		}
+
+		next, err = semver.BumpNextFunc(current)
+		if err != nil {
+			return fmt.Errorf("failed to determine next version: %w", err)
+		}
+	default:
+		return cli.Exit("invalid --label: must be 'patch', 'minor', or 'major'", 1)
+	}
+
+	switch {
+	case meta != "":
+		next.Build = meta
+	case isPreserveMeta:
+		next.Build = current.Build
+	default:
+		next.Build = ""
+	}
+
+	if err := semver.SaveVersion(path, next); err != nil {
+		return fmt.Errorf("failed to save version: %w", err)
+	}
+
+	fmt.Printf("Bumped version from %s to %s\n", current.String(), next.String())
+	return nil
+}
+
+// promotePreRelease strips pre-release and optionally preserves metadata.
+func promotePreRelease(current semver.SemVersion, preserveMeta bool) semver.SemVersion {
+	next := current
+	next.PreRelease = ""
+	if preserveMeta {
+		next.Build = current.Build
+	} else {
+		next.Build = ""
+	}
+	return next
+}
+
+// tryInferBumpTypeFromCommitParserPlugin tries to infer bump type from commit messages.
+func tryInferBumpTypeFromCommitParserPlugin(since, until string) string {
+	parser := commitparser.GetCommitParserFn()
+	if parser == nil {
+		return ""
+	}
+
+	commits, err := gitlog.GetCommitsFn(since, until)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "failed to read commits: %v\n", err)
+		return ""
+	}
+
+	label, err := parser.Parse(commits)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "commit parser failed: %v\n", err)
+		return ""
+	}
+
+	return label
+}
