+++ cmd/semver/setcmd/setcmd_test.go
@@ -1,0 +1,122 @@
+package setcmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/indaco/semver-cli/internal/config"
+	"github.com/indaco/semver-cli/internal/testutils"
+	"github.com/urfave/cli/v3"
+)
+
+func TestCLI_SetVersionCommandVariants(t *testing.T) {
+	tmpDir := t.TempDir()
+	versionPath := filepath.Join(tmpDir, ".version")
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	tests := []struct {
+		name     string
+		args     []string
+		expected string
+	}{
+		{"set version", []string{"semver", "set", "2.5.0"}, "2.5.0"},
+		{"set with pre-release", []string{"semver", "set", "3.0.0", "--pre", "beta.2"}, "3.0.0-beta.2"},
+		{"set with metadata", []string{"semver", "set", "1.0.0", "--meta", "001"}, "1.0.0+001"},
+		{"set with pre and meta", []string{"semver", "set", "1.0.0", "--pre", "alpha.1", "--meta", "exp.sha.5114f85"}, "1.0.0-alpha.1+exp.sha.5114f85"},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			testutils.RunCLITest(t, appCli, tt.args, tmpDir)
+			got := testutils.ReadTempVersionFile(t, tmpDir)
+			if got != tt.expected {
+				t.Errorf("expected %q, got %q", tt.expected, got)
+			}
+		})
+	}
+}
+
+func TestCLI_SetVersionCommand_InvalidFormat(t *testing.T) {
+	tmpDir := t.TempDir()
+	versionPath := filepath.Join(tmpDir, ".version")
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	err := appCli.Run(context.Background(), []string{"semver", "set", "invalid.version"})
+	if err == nil {
+		t.Fatal("expected error due to invalid version format, got nil")
+	}
+	if !strings.Contains(err.Error(), "invalid version format") {
+		t.Errorf("unexpected error: %v", err)
+	}
+}
+
+func TestCLI_SetVersionCommand_MissingArgument(t *testing.T) {
+	if os.Getenv("TEST_SEMVER_SET_MISSING_ARG") == "1" {
+		tmp := t.TempDir()
+		versionPath := filepath.Join(tmp, ".version")
+
+		// Prepare and run the CLI command
+		cfg := &config.Config{Path: versionPath}
+		appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+		err := appCli.Run(context.Background(), []string{"semver", "set", "--path", versionPath})
+		if err != nil {
+			fmt.Fprintln(os.Stderr, err)
+			os.Exit(1) // expected non-zero exit
+		}
+		os.Exit(0) // ❌ should not happen
+	}
+
+	cmd := exec.Command(os.Args[0], "-test.run=TestCLI_SetVersionCommand_MissingArgument")
+	cmd.Env = append(os.Environ(), "TEST_SEMVER_SET_MISSING_ARG=1")
+	output, err := cmd.CombinedOutput()
+
+	if err == nil {
+		t.Fatal("expected non-zero exit status")
+	}
+
+	expected := "missing required version argument"
+	if !strings.Contains(string(output), expected) {
+		t.Errorf("expected output to contain %q, got %q", expected, string(output))
+	}
+}
+
+func TestCLI_SetVersionCommand_SaveError(t *testing.T) {
+	tmp := t.TempDir()
+
+	protectedDir := filepath.Join(tmp, "protected")
+	if err := os.Mkdir(protectedDir, 0555); err != nil {
+		t.Fatal(err)
+	}
+	t.Cleanup(func() {
+		_ = os.Chmod(protectedDir, 0755)
+	})
+
+	versionPath := filepath.Join(protectedDir, ".version")
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	err := appCli.Run(context.Background(), []string{
+		"semver", "set", "3.0.0", "--path", versionPath,
+	})
+	if err == nil {
+		t.Fatal("expected error due to save failure, got nil")
+	}
+
+	if !strings.Contains(err.Error(), "failed to save version") {
+		t.Errorf("unexpected error message: %v", err)
+	}
+}
