+++ cmd/semver/showcmd/showcmd_test.go
@@ -1,0 +1,112 @@
+package showcmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/indaco/semver-cli/internal/config"
+	"github.com/indaco/semver-cli/internal/testutils"
+	"github.com/urfave/cli/v3"
+)
+
+func TestCLI_ShowCommand(t *testing.T) {
+	tmpDir := t.TempDir()
+	testutils.WriteTempVersionFile(t, tmpDir, "9.8.7")
+	versionPath := filepath.Join(tmpDir, ".version")
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, []string{"semver", "show"}, tmpDir)
+	})
+	if err != nil {
+		t.Fatalf("Failed to capture stdout: %v", err)
+	}
+
+	if output != "9.8.7" {
+		t.Errorf("expected output '9.8.7', got %q", output)
+	}
+}
+
+func TestCLI_ShowCommand_NoAutoInit_MissingFile(t *testing.T) {
+	if os.Getenv("TEST_SEMVER_NO_AUTO_INIT") == "1" {
+		tmp := t.TempDir()
+		versionPath := filepath.Join(tmp, ".version")
+
+		// Prepare and run the CLI command
+		cfg := &config.Config{Path: versionPath}
+		appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+		err := appCli.Run(context.Background(), []string{"semver", "show", "--no-auto-init"})
+		if err != nil {
+			fmt.Fprintln(os.Stderr, err)
+		}
+		return
+	}
+
+	cmd := exec.Command(os.Args[0], "-test.run=TestCLI_ShowCommand_NoAutoInit_MissingFile")
+	cmd.Env = append(os.Environ(), "TEST_SEMVER_NO_AUTO_INIT=1")
+	output, err := cmd.CombinedOutput()
+
+	if err == nil {
+		t.Fatal("expected non-zero exit status")
+	}
+
+	expected := "version file not found"
+	if !strings.Contains(string(output), expected) {
+		t.Errorf("expected output to contain %q, got %q", expected, string(output))
+	}
+}
+func TestCLI_ShowCommand_NoAutoInit_FileExists(t *testing.T) {
+	tmpDir := t.TempDir()
+	testutils.WriteTempVersionFile(t, tmpDir, "1.2.3")
+	versionPath := filepath.Join(tmpDir, ".version")
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	output, err := testutils.CaptureStdout(func() {
+		testutils.RunCLITest(t, appCli, []string{"semver", "show", "--no-auto-init"}, tmpDir)
+	})
+	if err != nil {
+		t.Fatalf("Failed to capture stdout: %v", err)
+	}
+
+	if output != "1.2.3" {
+		t.Errorf("expected output '1.2.3', got %q", output)
+	}
+}
+
+func TestCLI_ShowCommand_InvalidVersionContent(t *testing.T) {
+	tmp := t.TempDir()
+	versionPath := filepath.Join(tmp, ".version")
+
+	// Write an invalid version string
+	if err := os.WriteFile(versionPath, []byte("not-a-semver\n"), 0644); err != nil {
+		t.Fatalf("failed to write invalid version: %v", err)
+	}
+
+	// Prepare and run the CLI command
+	cfg := &config.Config{Path: versionPath}
+	appCli := testutils.BuildCLIForTests(cfg.Path, []*cli.Command{Run()})
+
+	err := appCli.Run(context.Background(), []string{"semver", "show"})
+	if err == nil {
+		t.Fatal("expected error due to invalid version, got nil")
+	}
+
+	if !strings.Contains(err.Error(), "failed to read version file at") {
+		t.Errorf("unexpected error message: %v", err)
+	}
+	if !strings.Contains(err.Error(), "invalid version format") {
+		t.Errorf("error does not mention 'invalid version format': %v", err)
+	}
+}
