+++ internal/clix/versionfile.go
@@ -1,0 +1,36 @@
+package clix
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/indaco/semver-cli/internal/semver"
+	"github.com/urfave/cli/v3"
+)
+
+// FromCommand extracts the --path and --no-auto-init flags from a cli.Command,
+// and passes them to GetOrInitVersionFile.
+func FromCommand(cmd *cli.Command) (bool, error) {
+	return GetOrInitVersionFile(cmd.String("path"), cmd.Bool("no-auto-init"))
+}
+
+// GetOrInitVersionFile initializes the version file at the given path
+// or checks for its existence based on the noAutoInit flag.
+// It returns true if the file was created, false if it already existed.
+func GetOrInitVersionFile(path string, noAutoInit bool) (bool, error) {
+	if noAutoInit {
+		if _, err := os.Stat(path); err != nil {
+			return false, cli.Exit(fmt.Sprintf("version file not found at %s", path), 1)
+		}
+		return false, nil
+	}
+
+	created, err := semver.InitializeVersionFileWithFeedback(path)
+	if err != nil {
+		return false, err
+	}
+	if created {
+		fmt.Printf("Auto-initialized %s with default version\n", path)
+	}
+	return created, nil
+}
